requirements.txt
Django==1.9.4
six==1.10.0
Pillow==3.1.1
django_cleanup==0.4.2
django-bootstrap3==7.0.1
django-markdownx==1.4.2